Rename SideBar to AppSideBar & Move AppSideBar.tsx from components/ to layout/
