+++ src/layout/AppSideBar.tsx
@@ -20,7 +20,7 @@ import {
   SidebarMenuButton,
   SidebarMenuItem,
 } from "@/components/ui/sidebar";
-import { Button } from "./ui/button";
+import { Button } from "../components/ui/button";
 import {
   ConnectModal,
   useWallets,
